+++ libs/frontend/modules/tag/src/graphql/TagGraph.fragment.graphql.gen.ts
@@ -3,11 +3,13 @@ import * as Types from '@codelab/frontend/abstract/codegen'
 import {
   TagFragment,
   TagEdgeFragment,
+  TagGraphFragment,
 } from '../use-cases/Tag.fragment.graphql.gen'
 import { gql } from '@apollo/client'
 import {
   TagFragmentDoc,
   TagEdgeFragmentDoc,
+  TagGraphFragmentDoc,
 } from '../use-cases/Tag.fragment.graphql.gen'
 export type TagGraphFragment = {
   vertices: Array<TagFragment>
