--- libs/frontend/modules/tag/src/use-cases/create-tag/CreateTag.web.graphql
@@ -1,5 +1,0 @@
-mutation CreateTag($input: CreateTagInput!) {
-  createTag(input: $input) {
-    id
-  }
-}
