--- libs/frontend/modules/tag/src/use-cases/delete-tags/DeleteTags.web.graphql
@@ -1,5 +1,0 @@
-mutation DeleteTags($input: DeleteTagsInput!) {
-  deleteTags(input: $input) {
-    ...Tag
-  }
-}
