--- libs/frontend/modules/tag/src/use-cases/export-tags/ExportTags.web.graphql
@@ -1,5 +1,0 @@
-query ExportTags($input: GetTagGraphsInput) {
-  getTagGraphs(input: $input) {
-    ...TagGraph
-  }
-}
