--- libs/frontend/modules/tag/src/use-cases/get-tag-graph/GetTagGraph.web.graphql
@@ -1,5 +1,0 @@
-query GetTagGraph {
-  getTagGraph {
-    ...TagGraph
-  }
-}
