--- libs/frontend/modules/tag/src/use-cases/get-tag-graphs/GetTagGraphs.web.graphql
@@ -1,5 +1,0 @@
-query GetTagGraphs {
-  getTagGraphs {
-    ...TagGraph
-  }
-}
