--- libs/frontend/modules/tag/src/use-cases/get-tags/GetTags.web.graphql
@@ -1,5 +1,0 @@
-query GetTags {
-  getTags {
-    ...Tag
-  }
-}
