--- libs/frontend/modules/tag/src/use-cases/import-tags/ImportTags.web.graphql
@@ -1,3 +1,0 @@
-mutation ImportTags($input: ImportTagsInput!) {
-  importTags(input: $input)
-}
