--- libs/frontend/modules/tag/src/use-cases/update-tag/UpdateTag.web.graphql
@@ -1,5 +1,0 @@
-mutation UpdateTag($input: UpdateTagInput!) {
-  updateTag(input: $input) {
-    ...Tag
-  }
-}
