--- libs/frontend/modules/tag/src/use-cases/update-tag/UpdateTag.web.graphql.gen.ts
@@ -1,63 +1,0 @@
-import * as Types from '@codelab/frontend/abstract/codegen'
-
-import { TagFragment, TagEdgeFragment } from '../Tag.fragment.graphql.gen'
-import { gql } from '@apollo/client'
-import { TagFragmentDoc, TagEdgeFragmentDoc } from '../Tag.fragment.graphql.gen'
-import * as Apollo from '@apollo/client'
-const defaultOptions = {}
-export type UpdateTagMutationVariables = Types.Exact<{
-  input: Types.UpdateTagInput
-}>
-
-export type UpdateTagMutation = { updateTag?: TagFragment | null | undefined }
-
-export const UpdateTagGql = gql`
-  mutation UpdateTag($input: UpdateTagInput!) {
-    updateTag(input: $input) {
-      ...Tag
-    }
-  }
-  ${TagFragmentDoc}
-`
-export type UpdateTagMutationFn = Apollo.MutationFunction<
-  UpdateTagMutation,
-  UpdateTagMutationVariables
->
-
-/**
- * __useUpdateTagMutation__
- *
- * To run a mutation, you first call `useUpdateTagMutation` within a React component and pass it any options that fit your needs.
- * When your component renders, `useUpdateTagMutation` returns a tuple that includes:
- * - A mutate function that you can call at any time to execute the mutation
- * - An object with fields that represent the current status of the mutation's execution
- *
- * @param baseOptions options that will be passed into the mutation, supported options are listed on: https://www.apollographql.com/docs/react/api/react-hooks/#options-2;
- *
- * @example
- * const [updateTagMutation, { data, loading, error }] = useUpdateTagMutation({
- *   variables: {
- *      input: // value for 'input'
- *   },
- * });
- */
-export function useUpdateTagMutation(
-  baseOptions?: Apollo.MutationHookOptions<
-    UpdateTagMutation,
-    UpdateTagMutationVariables
-  >,
-) {
-  const options = { ...defaultOptions, ...baseOptions }
-  return Apollo.useMutation<UpdateTagMutation, UpdateTagMutationVariables>(
-    UpdateTagGql,
-    options,
-  )
-}
-export type UpdateTagMutationHookResult = ReturnType<
-  typeof useUpdateTagMutation
->
-export type UpdateTagMutationResult = Apollo.MutationResult<UpdateTagMutation>
-export type UpdateTagMutationOptions = Apollo.BaseMutationOptions<
-  UpdateTagMutation,
-  UpdateTagMutationVariables
->
